refactor/cleanup/streamline widgets
