+++ src/views/default/tag.php
@@ -14,6 +14,6 @@ $this->title = Yii::$app->settings->get('site', 'publication.default-tag') ?: Yi
 ?>
 <div class="publication-default-tag publication-default-index">
     <?= Cell::widget(['id' => 'publication_tag_top', 'requestParam' => 'publication_tag_top']) ?>
-    <?= TaggedPublication::widget(['tagId' => $tagId,'categoryId' => $categoryId, 'limit' => $limit]) ?>
+    <?= TaggedPublication::widget(['tagId' => $tagId,'categoryId' => $categoryId, 'limit' => $limit,'pagination' => true]) ?>
     <?= Cell::widget(['id' => 'publication_tag_bottom', 'requestParam' => 'publication_tag_bottom']) ?>
 </div>
